pyproject.toml
[project]
name = "healthie-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Brandon Redmond", email = "brandon@gethealthie.om" }
]
requires-python = ">=3.13"
dependencies = [
    "graphql-core>=3.2.6",
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "PyYAML>=6.0.1",
]

[project.scripts]
healthie-mcp = "healthie_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=6.2.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[tool.pytest.ini_options]
# Test discovery
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test execution
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=src/healthie_mcp",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=85",
]

# Async testing
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Test markers
markers = [
    "unit: Unit tests that test individual components in isolation",
    "integration: Integration tests that test component interactions",
    "e2e: End-to-end tests that test complete workflows",
    "slow: Tests that take longer than 1 second to run",
    "requires_api: Tests that require access to external APIs",
    "requires_auth: Tests that require valid authentication credentials",
]

# Filtering
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# Minimum Python version for tests
minversion = "6.0"
